Separate between Work Item Type Field Type String and Markup (#477)
Added a `system.description.markup` to describe the type
of markup used in the associated `system.description` field.
Also added some possible values as constants, which are used
when importing remote work items from GitHub (`Markdown`) and
JIRA (`JiraWiki`).
Also, when a new Work Item is created, the markup type associated
with the description is 'Asciidoc'.

Note that this PR does not deal with the HTML rendering of the
description.

Fixes #477

Signed-off-by: Xavier Coulon <xcoulon@redhat.com>